requirement.txt
dataclasses==0.6
Django==2.1.4
djangorestframework==3.9.0
djongo==1.2.31
PyJWT==1.7.1
pymongo==3.7.2
pytz==2018.7
sqlparse==0.2.4
